refine page content, correct a couple of mistakes
I hope you may agree with all of these. I had tried to do it in 3 commits, each one you could accept alone (the first was a needed correction to the initial admin URL, to add /ghost to the end, the second was refinements to the discussion of removing the initial user and posts,  and then the 3rd was a number of grammatical improvement suggestions. I realize some folks don't care to bother with the latter, which is why I wanted to offer these in 3 commits.) 

But I am having trouble figuring how to do that with the github web UI.  I've usually only ever proposed one set of changes at a time. Apologies for coming up lame, in offering all these changes at once.